增加下载状态、修改数据表数据类型

+++ src/main/java/com/devteam/mikufunbackend/dao/ResourceInformationDao.java
@@ -55,7 +55,7 @@ public interface ResourceInformationDao {
      * @return
      */
     int updatePlayPosition(int fileId,
-                           int recentPlayPosition);
+                           double recentPlayPosition);
 
     /**
      * 更新视频最近播放时间
